Moved the social link section into it's own view
